Extract SecondaryPane into VenueDetailsPane.tsx
